Select proxy values

Select GDP or PCE as proxy value based on pre-specified rules.

Usage

select_proxy(svy_table, region_code, ref_gdp, ref_pce)

Arguments

svy_table

data.frame: A table with information on surveys and national accounts data.

region_code

character: World Bank three letter region code.

ref_gdp

numeric: GDP value for the reference year.

ref_pce

numeric: PCE value for the reference year.

Value

list

Details

PCE is the default for all countries, except in Sub-Saharan Africa where GDP is always used. If any PCE value is missing for the specific reference year, GDP is used.

References

Prydz, E.B., D. Jolliffe, C. Lakner, D.G. Mahler, P. Sangraula. 2019. "National Accounts Data used in Global Poverty Measurement". Global Poverty Monitoring Technical Note 8. World Bank, Washington, DC.